xen: align ram bank 0 configuration with libxl
Align GUEST_RAM0_SIZE with value in libxl.

Ensure that the first extended region does not overlap with
devices memory addresses.

Changes have been synced with the 'xen-4.19-xt0.1' branch
on the https://github.com/xen-troops/xen.git repository.

Note: this fix is not upstreamable. It is necessary to implement
      a proper fix while considering the memory ranges used in
      the DTS.

Signed-off-by: Mykola Kvach <mykola_kvach@epam.com>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Acked-by: Dmytro Firsov <dmytro_firsov@epam.com>